finalizacion

+++ src/users/users.service.ts
@@ -1,15 +1,31 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './user.entity';
 import { Repository } from 'typeorm';
 import { CreateUserDTO } from './dto/create-user.dto';
+import { UpdateUserDTO } from './dto/update-user.dto';
+import { CreateProfileDto } from './dto/create-profile.dto';
+import { Profile } from './profile.entity';
 
 @Injectable()
 export class UsersService {
 
-  constructor(@InjectRepository(User) private userRepository: Repository<User>) { }
+  constructor(
+    @InjectRepository(User) private userRepository: Repository<User>,
+    @InjectRepository(Profile) private profileRepository: Repository<Profile>
+  ) { }
 
-  createUser(user: CreateUserDTO) {
+  async createUser(user: CreateUserDTO) {
+    const userFound = await this.userRepository.findOne({
+      where: {
+        username: user.username
+      }
+    })
+
+    if (userFound) {
+      return new HttpException('User already exists', HttpStatus.CONFLICT);
+    }
+
     const newUser = this.userRepository.create(user);
     return this.userRepository.save(newUser)
   }
@@ -18,15 +34,58 @@ export class UsersService {
     return this.userRepository.find();
   }
 
-  getUser(id: number) {
-    return this.userRepository.findOne({
+  async getUser(id: number) {
+    const userFound = await this.userRepository.findOne({
+      where: {
+        id
+      },
+      relations: ['posts', 'profile']
+    })
+
+    if (!userFound) {
+      return new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+    return userFound;
+  }
+
+  async deleteUser(id: number) {
+    const result = await this.userRepository.delete({ id });
+
+    if (result.affected === 0) {
+      return new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+
+    return this.userRepository.delete({ id })
+  }
+
+  async updateUser(id: number, user: UpdateUserDTO) {
+    const userFound = await this.userRepository.findOne({
       where: {
         id
       }
-    })
+    });
+
+    if (!userFound) {
+      return new HttpException('user not found', HttpStatus.NOT_FOUND);
+    }
+
+    const updateUser = Object.assign(userFound, user);
+    return this.userRepository.save(updateUser);
   }
 
-  deleteUser(id: number) {
-    return this.userRepository.delete({ id});
+  async createProfile(id: number, profile: CreateProfileDto) {
+    const userFound = await this.userRepository.findOne({ where: { id } });
+
+    if (!userFound) {
+      return new HttpException('User not Found', HttpStatus.NOT_FOUND);
+    }
+
+    const newProfile = this.profileRepository.create(profile);
+
+    const savedProfile = await this.profileRepository.save(newProfile);
+
+    userFound.profile = savedProfile;
+
+    return this.userRepository.save(userFound);
   }
 }
